fix: keep upid_set_cfg() in single precision
The integral re-alignment in upid_set_cfg()
was done in double. On a single-precision FPU
(Cortex-M4F) that pulled in soft-float
helpers.

Changes:
- adjusted integral -> float, overflow caught
  with is_finite()
- <float.h> -> no longer needed, dropped
  (README "Needs" updated)
- example printf -> explicit (double) cast,
  clean under -Wdouble-promotion

Cortex-M4F -Os: 1440 -> 1336 B text, no
__aeabi_d* references left. When the re-clamp
moves the output the integral may differ by
1 ULP from the double version; otherwise it
is bit-identical.

+++ examples/basic_pid.c
@@ -64,7 +64,7 @@ int main(void)
 
 		if (step % 200 == 0)
 			printf("t=%5.2fs  measurement=%7.3f  output=%7.3f\n",
-			       step * dt_s, (double)read_measurement(),
+			       (double)(step * dt_s), (double)read_measurement(),
 			       (double)output);
 	}
